refactor: merge tests into single module
This is required for running `swift test`. This adds a flag
`-enable-testing` which exposes the module's internal API for testing,
but thereby prevents the functions from being instantiated at their use
site and being specialised at a concrete type, which is breaking
compilation. Adding all functions to a single module and marking them as
private prevents this behaviour.